aira-voice-agent/requirements.txt
# ═══ FILE: requirements.txt ═══
# Purpose: List all Python dependencies for Aira Voice Agent
# Inputs: Standard Python environment
# Outputs: Installed packages for STT, LLM, TTS, and Backend

fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
openai-whisper==20231117
ollama==0.1.6
TTS==0.22.0
pyttsx3==2.90
pydub==0.25.1
python-dotenv==1.0.0
psutil==5.9.8
soundfile==0.12.1
numpy==1.26.3
jinja2==3.1.3
# torch and torchvision are required by TTS but should be installed manually for CPU/GPU optimization on Windows
